Lambda function process job data, retrieve instance-to-node mapping from OpenSearch, push job data to OpenSearch (#5507)

+++ api/client/src/pcluster_client/api/cluster_logs_api.py
@@ -99,7 +99,7 @@ class ClusterLogsApi(object):
                     'start_from_head':
                         (bool,),
                     'limit':
-                        (float,),
+                        (int,),
                     'start_time':
                         (datetime,),
                     'end_time':
@@ -306,7 +306,7 @@ class ClusterLogsApi(object):
             region (str): AWS Region that the operation corresponds to.. [optional]
             next_token (str): Token to use for paginated requests.. [optional]
             start_from_head (bool): If the value is true, the earliest log events are returned first. If the value is false, the latest log events are returned first. (Defaults to 'false'.). [optional]
-            limit (float): The maximum number of log events returned. If you don't specify a value, the maximum is as many log events as can fit in a response size of 1 MB, up to 10,000 log events.. [optional]
+            limit (int): The maximum number of log events returned. If you don't specify a value, the maximum is as many log events as can fit in a response size of 1 MB, up to 10,000 log events.. [optional]
             start_time (datetime): The start of the time range, expressed in ISO 8601 format (e.g. '2021-01-01T20:00:00Z'). Events with a timestamp equal to this time or later than this time are included.. [optional]
             end_time (datetime): The end of the time range, expressed in ISO 8601 format (e.g. '2021-01-01T20:00:00Z'). Events with a timestamp equal to or later than this time are not included.. [optional]
             _return_http_data_only (bool): response data without head status
